LOOK-357 refactoring for DAGs & added new tables
