+++ python/to_stage.py
@@ -1,18 +1,14 @@
 from airflow.models import Variable
-from airflow.providers.amazon.aws.hooks.redshift_sql import RedshiftSQLHook
 import logging
 import pendulum
 import json
 from python.core.configs import get_job_config
 from python.core.utils import get_fields_for
+from python.core.connections import redshift_conn_dev
 
 
-# getting connection to Redshift DB
-redshift_hook = RedshiftSQLHook(
-    redshift_conn_id='redshift_default',
-    schema='dev'
-)
-redshift_conn = redshift_hook.get_conn()
+# Get connection to Redshift DB
+redshift_conn = redshift_conn_dev()
 cursor = redshift_conn.cursor()
 
 
@@ -62,7 +58,7 @@ def stg_load(*op_args, **kwargs):
     pk = job_cfg.pk
     source_fields = get_fields_for('source', job_cfg.map)
     # target_fields = get_fields_for('target', job_cfg.map)
-    
+    custom_query = job_cfg.custom_query if job_cfg.custom_query is not None else None
 
     # check if table not exists in target schema
     query = f'''
@@ -141,16 +137,19 @@ def stg_load(*op_args, **kwargs):
             logging.info(f'{cursor.rowcount} rows inserted for {comp_id} at {pendulum.now()}')
         elif load_type == 'increment_with_delete':
             # creating temp table with new data increment
-            query = f'''
-                CREATE temporary TABLE {table}_{comp_id}_temp as
-                SELECT {source_fields}
-                FROM {ext_schema}.{table}
-                WHERE {increment} > (
-                    SELECT coalesce(max({increment}), '1970-01-01 00:00:00'::timestamp)
-                    FROM {target_schema}.{table}
-                    WHERE comp_id = {comp_id}
-                ) and {increment} < CURRENT_DATE + interval '8 hours'
-            '''
+            if custom_query is not None:
+                query = custom_query.format(table=table, comp_id=comp_id, ext_schema=ext_schema, increment=increment, target_schema=target_schema)
+            else:
+                query = f'''
+                    CREATE temporary TABLE {table}_{comp_id}_temp as
+                    SELECT {source_fields}
+                    FROM {ext_schema}.{table}
+                    WHERE {increment} > (
+                        SELECT coalesce(max({increment}), '1970-01-01 00:00:00'::timestamp)
+                        FROM {target_schema}.{table}
+                        WHERE comp_id = {comp_id}
+                    ) and {increment} < CURRENT_DATE + interval '8 hours'
+                '''
             cursor.execute(query)
             logging.info(f'Temp table is created')
             # deleting from target table data that were updated
